engine: introduce subsystem module layout
