+++ crates/izwi-core/src/engine/mod.rs
@@ -24,6 +24,7 @@
 //! └─────────────────────────────────────────────────────────────────┘
 //! ```
 
+mod cache;
 mod config;
 mod core;
 mod executor;
